+++ UtinyRipperCore/Parser/Classes/AnimatorController/State/StateRange.cs
@@ -4,11 +4,11 @@
 	{
 		public void Read(AssetStream stream)
 		{
-			StartIndex = stream.ReadUInt32();
-			Count = stream.ReadUInt32();
+			StartIndex = (int)stream.ReadUInt32();
+			Count = (int)stream.ReadUInt32();
 		}
 
-		public uint StartIndex { get; private set; }
-		public uint Count { get; private set; }
+		public int StartIndex { get; private set; }
+		public int Count { get; private set; }
 	}
 }
